pyproject.toml
[build-system]
  build-backend = "uv_build"
  requires = ["uv_build"]

[dependency-groups]
  dev = ["dycw-utilities[test]>=0.184.4", "pyright>=1.1.408", "rich>=14.2.0"]

[project]
  dependencies = [
    "click>=8.3.1",
    "dycw-pre-commit-hooks>=0.14.78",
    "dycw-utilities>=0.184.4",
  ]
  description = "Pre-commit hooks"
  name = "qrt-pre-commit-hooks"
  readme = "README.md"
  requires-python = ">= 3.13"
  version = "0.3.0"

  [project.scripts]
    modify-ci-pull-request = "qrt_pre_commit_hooks.hooks.modify_ci_pull_request:_main"
    modify-ci-push = "qrt_pre_commit_hooks.hooks.modify_ci_push:_main"
    modify-pre-commit = "qrt_pre_commit_hooks.hooks.modify_pre_commit:_main"
    modify-pyproject = "qrt_pre_commit_hooks.hooks.modify_pyproject:_main"

  [project.optional-dependencies]
    cli = [
      "click==8.3.1",
      "dycw-pre-commit-hooks==0.14.78",
      "dycw-utilities==0.184.4",
    ]

[tool]
  [tool.uv]
    [tool.uv.build-backend]
      module-name = "qrt_pre_commit_hooks"
      module-root = "src"

    [[tool.uv.index]]
      explicit = true
      name = "gitea"
      url = "https://qrt-bot:e43d1df41a3ecf96e4adbaf04e98cfaf094d253e@gitea.main:3000/api/packages/qrt/pypi/simple"
